go/druid: add multiexporter package
